fix: 다이어리 삭제 시 이미지 `null` 체크

+++ src/main/java/picdiary/diary/controller/DiaryController.java
@@ -89,8 +89,10 @@ public class DiaryController {
     @DeleteMapping("/{diaryId}")
     public ApplicationResponse<Long> deleteDiary(@Parameter(hidden = true) UserEntity user, @PathVariable("diaryId") Long diaryId) {
         DiaryEntity diary = diaryService.findDiaryById(diaryId);
-        s3Service.deleteImage(diary.getImageFileName());
-        diaryService.deleteDiary(user.getId(), diaryId);
+        if (diary.getImageFileName() != null) {
+            s3Service.deleteImage(diary.getImageFileName());
+        }
+        diaryService.deleteDiary(user.getId(), diary);
         return ApplicationResponse.success(diaryId, "다이어리가 삭제되었습니다.");
     }
 }
